Refactor: group components into containers folder
